+++ run.py
@@ -1,10 +1,10 @@
 ''''this is the entry point to fire up the app'''
 import os
 
-from app import create_app
+from app.api.v1 import create_app
 
 config = os.getenv('APP_SETTINGS')
 app = create_app(config)
 
 if __name__ == '__main__':
-	app.run(debug=True, port=5000)
+	app.run(debug=True, port=5003)
